First pass at adding gradients to background Cleaned up Settings page

+++ KoalaMama/KoalaMama/HMIntroFourViewController.m
@@ -62,7 +62,7 @@
     Mother* mother = [Mother getMother];
     NSNumberFormatter * f = [[NSNumberFormatter alloc] init];
     mother.imperialPrePregnancyWeight = [f numberFromString:[self.prePregnancyWeightInput text]];
-    mother.imperialHeight = [NSNumber numberWithInt:(([self.heightPicker selectedRowInComponent:0] + 4) * 12 + (int)[self.heightPicker selectedRowInComponent:1])];
+    mother.imperialHeight = [NSNumber numberWithInt:((int)([self.heightPicker selectedRowInComponent:0] + 4) * 12 + (int)[self.heightPicker selectedRowInComponent:1])];
 
     
     // Remember that user has now seen the tutorial
@@ -90,9 +90,9 @@
 
 - (NSString *)pickerView:(UIPickerView *)pickerView titleForRow:(NSInteger)row forComponent:(NSInteger)component {
     if (component == 0) {
-        return [NSString stringWithFormat:@"%d ft", row + 4];
+        return [NSString stringWithFormat:@"%d ft", (int)row + 4];
     } else {
-        return [NSString stringWithFormat:@"%d in", row];
+        return [NSString stringWithFormat:@"%d in", (int)row];
     }
 }
